Add functionality to load a given number of frames
The torch dataloader expects all tensores to have the same dimensions. Therefore you can set the number of frames to load from each video. There are different strategies to pad the frames if the video is too short or to select frames from the video if it is too long.